pyproject.toml
[tool.poetry]
name = "restproject"
version = "0.1.0"
description = ""
authors = ["annett21 <annettarasevich@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
celery = "5.2.7"
Django = "4.1.4"
django-rest-swagger = "2.2.0"
djangorestframework = "3.14.0"
djangorestframework-jwt = "1.11.0"
drf-yasg = "1.21.4"
drf-yasg2 = "1.19.1"
factory-boy = "3.2.1"
redis = "4.4.2"
time-machine = "2.9.0"
django-celery-beat = "2.4.0"
django-celery-results = "2.4.0"
aiogram = "2.25.1"
python-dotenv = "^1.0.0"
pytest-django = "^4.5.2"
psycopg2-binary = "^2.9.6"


[tool.poetry.group.dev.dependencies]
ptpython = "^3.0.22"
black = "^22.12.0"
flake8 = "^6.0.0"
django-extensions = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
